Effectful.Environment.Guard.Static

Effect

data EnvGuard (a :: Type -> Type) b

Static effect for EnvGuard.

Since: 0.1

Handler

runEnvGuard :: forall (es :: [Effect]) a. IOE :> es => Eff (EnvGuard ': es) a -> Eff es a

Runs an EnvGuard effect.

Since: 0.1

High level combinators

data ExpectEnv

The expectation for an environment variable lookup.

Since: env-guard-0.1.1

Constructors

ExpectEnvSet
Expect that the environment variable is set (i.e. contents can be anything).

Since: env-guard-0.1.1

ExpectEnvEquals String
Expect that the environment variable is set and the contents equals the string. This is case-insensitive.

Since: env-guard-0.1.1

ExpectEnvPredicate (String -> Bool)
Expect that the environment variable is set and its contents satisfies the predicate.

Since: env-guard-0.1.1

withGuard :: forall (es :: [Effect]) a. (EnvGuard :> es, HasCallStack) => String -> ExpectEnv -> Eff es a -> Eff es (Maybe a)

Since: 0.1

withGuard_ :: forall (es :: [Effect]) a. (EnvGuard :> es, HasCallStack) => String -> ExpectEnv -> Eff es a -> Eff es ()

Since: 0.1

guardOrElse

Arguments

:: forall (es :: [Effect]) a e. (EnvGuard :> es, HasCallStack)

=> String
The environment variable.

-> ExpectEnv
The expectation.

-> Eff es a
The action to run if the expectation succeeds.

-> Eff es e
The action to run if the expectation fails.

-> Eff es (Either e a)
The result.

Since: 0.1

guardOrElse'

Arguments

:: forall (es :: [Effect]) a. (EnvGuard :> es, HasCallStack)

=> String
The environment variable.

-> ExpectEnv
The expectation.

-> Eff es a
The action to run if the expectation succeeds.

-> Eff es a
The action to run if the expectation fails.

-> Eff es a
The result.

Since: 0.1

Low level functions

Checking environment variable is set

guardSet :: forall (es :: [Effect]) a. (EnvGuard :> es, HasCallStack) => String -> Eff es a -> Eff es (Maybe a)

Since: 0.1

guardSet_ :: forall (es :: [Effect]) a. (EnvGuard :> es, HasCallStack) => String -> Eff es a -> Eff es ()

Since: 0.1

Checking environment variable match

guardEquals :: forall (es :: [Effect]) a. (EnvGuard :> es, HasCallStack) => String -> String -> Eff es a -> Eff es (Maybe a)

Since: 0.1

guardEquals_ :: forall (es :: [Effect]) a. (EnvGuard :> es, HasCallStack) => String -> String -> Eff es a -> Eff es ()

Since: 0.1

Checking environment variable predicate

guardPredicate :: forall (es :: [Effect]) a. (EnvGuard :> es, HasCallStack) => String -> (String -> Bool) -> Eff es a -> Eff es (Maybe a)

Since: 0.1

guardPredicate_ :: forall (es :: [Effect]) a. (EnvGuard :> es, HasCallStack) => String -> (String -> Bool) -> Eff es a -> Eff es ()

Since: 0.1