Effectful.Environment.Static

Description

Static effects for System.Environment. The primary interface exists at: https://hackage.haskell.org/package/effectful-2.2.2.0/docs/Effectful-Environment.html.

This module exists to provide a static version of the Effectful.Environment.Dynamic, and to provide an Eff version of executablePath for base >= 4.17.

Since: 0.1

Effect

data Environment (a :: Type -> Type) b

An effect for querying and modifying the system environment.

Functions

getArgs :: forall (es :: [Effect]). Environment :> es => Eff es [String]

Lifted getArgs.

getProgName :: forall (es :: [Effect]). Environment :> es => Eff es String

Lifted getProgName.

executablePath :: forall (es :: [Effect]). (Environment :> es, HasCallStack) => Eff es QueryExePath

getExecutablePath :: forall (es :: [Effect]). Environment :> es => Eff es FilePath

Lifted getExecutablePath.

getEnv :: forall (es :: [Effect]). Environment :> es => String -> Eff es String

Lifted getEnv.

lookupEnv :: forall (es :: [Effect]). Environment :> es => String -> Eff es (Maybe String)

Lifted lookupEnv.

setEnv :: forall (es :: [Effect]). Environment :> es => String -> String -> Eff es ()

Lifted setEnv.

unsetEnv :: forall (es :: [Effect]). Environment :> es => String -> Eff es ()

Lifted unsetEnv.

withArgs :: forall (es :: [Effect]) a. Environment :> es => [String] -> Eff es a -> Eff es a

Lifted withArgs.

withProgName :: forall (es :: [Effect]) a. Environment :> es => String -> Eff es a -> Eff es a

Lifted withProgName.

getEnvironment :: forall (es :: [Effect]). Environment :> es => Eff es [(String, String)]

Lifted getEnvironment.

Handlers

runEnvironment :: forall (es :: [Effect]) a. (HasCallStack, IOE :> es) => Eff (Environment ': es) a -> Eff es a

Run the Environment effect.

Types

data QueryExePath

Result of querying for the executable path.

Since: 0.1

Constructors

NoQuery
If the system does not provide a reliable way to determine the current executable.

Since: 0.1

QueryResult (Maybe FilePath)
The result of querying the executable name.

Since: 0.1