Effectful.FileSystem.Handle.Internal

Typed Handle

newtype Handle (p :: HandleMode)

Wrapper for file handles that includes an index for the mode.

Since: 0.1

Constructors

MkHandle Handle

unHandle :: forall (p :: HandleMode). Handle p -> Handle

Since: 0.1

data HandleMode

Possible handle modes.

Since: 0.1

Constructors

HandleModeRead
Since: 0.1

HandleModeWrite
Since: 0.1

HandleModeReadWrite
Since: 0.1

appendToMode :: Bool -> IOMode

Type families

type family CanRead (hm :: HandleMode) where ...

Since: 0.1

Equations

CanRead 'HandleModeRead = ()

CanRead 'HandleModeWrite = Unsatisfiable ('Text "HandleModeWrite does not have Read permission.")

CanRead 'HandleModeReadWrite = ()

type family CanWrite (hm :: HandleMode) where ...

Since: 0.1

Equations

CanWrite 'HandleModeRead = Unsatisfiable ('Text "HandleModeRead does not have Write permission.")

CanWrite 'HandleModeWrite = ()

CanWrite 'HandleModeReadWrite = ()

Locks

newtype LockedHandle (p :: HandleMode)

File handle with lock.

Since: 0.1

Constructors

MkLockedHandle (Handle p)

unLockedHandle :: forall (p :: HandleMode). LockedHandle p -> Handle p

Since: 0.1

liftLocked :: forall (p :: HandleMode) a. HasCallStack => (HasCallStack => Handle p -> a) -> LockedHandle p -> a

Lifts a function on handle to one on locked handles.

Since: 0.1

Helpers

liftLock :: forall f (p :: HandleMode). Functor f => (Handle p -> f ()) -> Handle p -> f (LockedHandle p)

liftTryLock :: forall f (p :: HandleMode). Functor f => (Handle p -> f Bool) -> Handle p -> f (Maybe (LockedHandle p))

liftUnlock :: forall f (p :: HandleMode). Functor f => (Handle p -> f ()) -> LockedHandle p -> f (Handle p)

withLockedFile

Arguments

:: forall (p :: HandleMode) (es :: [Effect]) a. HasCallStack

=> (Handle p -> Eff es ())
Lock function.

-> (Handle p -> Eff es ())
Unlock function.

-> Handle p
Handle to lock.

-> (LockedHandle p -> Eff es a)
Callback with locked handle.

-> Eff es a

Runs a computation with a locked file.

withTryLockedFile

Arguments

:: forall (p :: HandleMode) a (es :: [Effect]). HasCallStack

=> (Handle p -> Eff es Bool)
Lock function.

-> (Handle p -> Eff es ())
Unlock function.

-> Handle p
Handle to lock.

-> (LockedHandle p -> Eff es a)
Handle callback.

-> Eff es (Maybe a)

Like withSharedLockedFile, except the lock attempt does not block.

Since: 0.1