Effectful.FileSystem.HandleRW.Static

Effect

data HandleRW (a :: Type -> Type) b

Static effect for reading a handle.

Since: 0.1

openBinaryFile :: forall (es :: [Effect]). (HandleRW :> es, HasCallStack) => OsPath -> Eff es (Handle 'HandleModeReadWrite)

Opens a file in ReadWrite mode.

Since: 0.1

withBinaryFile :: forall (es :: [Effect]) a. (HandleRW :> es, HasCallStack) => OsPath -> (Handle 'HandleModeReadWrite -> Eff es a) -> Eff es a

Opens a file in ReadWrite mode.

Since: 0.1

Handlers

runHandleRW :: forall (es :: [Effect]) a. (HasCallStack, IOE :> es) => Eff (HandleRW ': es) a -> Eff es a

Runs HandleRW in IO.

Since: 0.1

Reexports

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Char8 it can be interpreted as containing 8-bit characters.

data Handle (p :: HandleMode)

Wrapper for file handles that includes an index for the mode.

Since: 0.1

data HandleMode

Possible handle modes.

Since: 0.1

Constructors

HandleModeReadWrite
Since: 0.1

type OsPath = OsString

Type representing filenames/pathnames.

This type doesn't add any guarantees over OsString.