Effectful.FileSystem.HandleReader.Dynamic

Description

Provides a dynamic effect for reading a handle.

Since: 0.1

Effect

data HandleReader (a :: Type -> Type) b where

Dynamic effect for reading a handle.

Since: 0.1

Constructors

OpenBinaryFile :: forall (a :: Type -> Type). OsPath -> HandleReader a (Handle 'HandleModeRead)

WithBinaryFile :: forall (a :: Type -> Type) b. OsPath -> (Handle 'HandleModeRead -> a b) -> HandleReader a b

HIsEOF :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a Bool

HGetBuffering :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a BufferMode

HIsOpen :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a Bool

HIsClosed :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a Bool

HIsReadable :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a Bool

HIsWritable :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a Bool

HIsSeekable :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a Bool

HIsTerminalDevice :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a Bool

HGetEcho :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a Bool

HWaitForInput :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> Int -> HandleReader a Bool

HReady :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a Bool

HGetChar :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a Char

HGetLine :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a ByteString

HGetContents :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a ByteString

HGet :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> Int -> HandleReader a ByteString

HGetSome :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> Int -> HandleReader a ByteString

HGetNonBlocking :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> Int -> HandleReader a ByteString

HLockRaw :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a ()

HTryLockRaw :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a Bool

HUnlockRaw :: forall (p :: HandleMode) (a :: Type -> Type). CanRead p => Handle p -> HandleReader a ()

openBinaryFile :: forall (es :: [Effect]). (HandleReader :> es, HasCallStack) => OsPath -> Eff es (Handle 'HandleModeRead)

Lifted openBinaryFile.

Since: 0.1

withBinaryFile :: forall (es :: [Effect]) a. (HandleReader :> es, HasCallStack) => OsPath -> (Handle 'HandleModeRead -> Eff es a) -> Eff es a

Lifted withBinaryFile.

Since: 0.1

hIsEOF :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Bool

Lifted hIsEof.

Since: 0.1

hGetBuffering :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es BufferMode

Lifted hGetBuffering.

Since: 0.1

hIsOpen :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Bool

Lifted hIsOpen.

Since: 0.1

hIsClosed :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Bool

Lifted hIsClosed.

Since: 0.1

hIsReadable :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Bool

Lifted hIsReadable.

Since: 0.1

hIsWritable :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Bool

Lifted hIsWritable.

Since: 0.1

hIsSeekable :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Bool

Lifted hIsSeekable.

Since: 0.1

hIsTerminalDevice :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Bool

Lifted hIsTerminalDevice.

Since: 0.1

hGetEcho :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Bool

Lifted hGetEcho.

Since: 0.1

hWaitForInput :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Int -> Eff es Bool

Lifted hWaitForInput.

Since: 0.1

hReady :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Bool

Lifted hReady.

Since: 0.1

hGetChar :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Char

Lifted hGetChar.

Since: 0.1

hGetLine :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es ByteString

Lifted hGetLine.

Since: 0.1

hGetContents :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es ByteString

Lifted hGetContents.

Since: 0.1

hGet :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Int -> Eff es ByteString

Lifted hGet.

Since: 0.1

hGetSome :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Int -> Eff es ByteString

Lifted hGetSome.

Since: 0.1

hGetNonBlocking :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Int -> Eff es ByteString

Lifted hGetNonBlocking.

Since: 0.1

hLockRaw :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es ()

Attempts to shared lock a file, blocking or throwing an exception upon failure.

Since: 0.1

hTryLockRaw :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Bool

Attempts to shared lock a file.

Since: 0.1

hUnlockRaw :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es ()

Unlocks a locked file.

Since: 0.1

Handlers

runHandleReader :: forall (es :: [Effect]) a. (HasCallStack, IOE :> es) => Eff (HandleReader ': es) a -> Eff es a

Runs HandleReader in IO.

Since: 0.1

File handles

data Handle (p :: HandleMode)

Wrapper for file handles that includes an index for the mode.

Since: 0.1

data HandleMode

Possible handle modes.

Since: 0.1

Constructors

HandleModeRead
Since: 0.1

type family CanRead (hm :: HandleMode) where ...

Since: 0.1

Equations

CanRead 'HandleModeRead = ()

CanRead 'HandleModeWrite = Unsatisfiable ('Text "HandleModeWrite does not have Read permission.")

CanRead 'HandleModeReadWrite = ()

Locking

These functions bring some type-safety to file locking. Consider:

main :: (HandleReader :> es) => Eff es ()
main = withBinaryFile path False $ \h -> do
  hLockRaw handle
  bs <- readBytes @HandleModeRead h
  hUnlockRaw handle
  print bs

readBytes :: (CanRead p, HandleReader :> es) => Handle p -> Eff es ByteString
readBytes handle = hGet handle 1024

In this example, we could remove all locking logic from main and everything would still compile. On the other hand:

main :: (HandleReader :> es) => Eff es ()
main = withBinaryFile path False $ \h -> withLockedFile h $ \lh -> do
  bs <- readBytes @HandleModeRead lh
  print bs

readBytes :: (CanRead p, HandleReader :> es) => LockedHandle p -> Eff es ByteString
readBytes lockedHandle = liftLocked (\h -> hGet h 1024) lockedHandle

Removing withLockedFile would cause a compilation error, since writeBytes requires a LockedHandle. The idea is to write most of the program's logic in terms of LockedHandle, using liftLocked to lift Handle functions.

data LockedHandle (p :: HandleMode)

File handle with lock.

Since: 0.1

liftLocked :: forall (p :: HandleMode) a. HasCallStack => (HasCallStack => Handle p -> a) -> LockedHandle p -> a

Lifts a function on handle to one on locked handles.

Since: 0.1

withLockedFile

Arguments

:: forall (p :: HandleMode) (es :: [Effect]) a. (CanRead p, HasCallStack, HandleReader :> es)

=> Handle p
Handle to lock.

-> (LockedHandle p -> Eff es a)
Callback with locked handle.

-> Eff es a

Runs a computation with a shared locked file.

Since: 0.1

withTryLockedFile

Arguments

:: forall (p :: HandleMode) a (es :: [Effect]). (CanRead p, HasCallStack, HandleReader :> es)

=> Handle p
Handle to lock.

-> (LockedHandle p -> Eff es a)
Handle callback.

-> Eff es (Maybe a)

Like withSharedLockedFile, except the lock attempt does not block.

Since: 0.1

hLock

Arguments

:: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HasCallStack, HandleReader :> es)

=> Handle p
Handle to lock.

-> Eff es (LockedHandle p)

Like hLockRaw, but returns a LockedHandle.

Since: 0.1

hTryLock

Arguments

:: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HasCallStack, HandleReader :> es)

=> Handle p
Handle to lock.

-> Eff es (Maybe (LockedHandle p))

Like hTryLockRaw, but returns a LockedHandle if it succeeds.

Since: 0.1

hUnlock

Arguments

:: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HasCallStack, HandleReader :> es)

=> LockedHandle p
Handle to unlock.

-> Eff es (Handle p)

Like hUnlockRaw, but returns the original handle.

Since: 0.1

Raw

withLockedFileRaw

Arguments

:: forall (p :: HandleMode) (es :: [Effect]) a. (CanRead p, HasCallStack, HandleReader :> es)

=> Handle p
Handle to lock.

-> Eff es a
Callback with locked handle.

-> Eff es a

withLockedFile without LockedHandle.

Since: 0.1

withTryLockedFileRaw

Arguments

:: forall (p :: HandleMode) a (es :: [Effect]). (CanRead p, HasCallStack, HandleReader :> es)

=> Handle p
Handle to lock.

-> Eff es a
Handle callback.

-> Eff es (Maybe a)

withTryLockedFileRaw without LockedHandle.

Since: 0.1

UTF-8 Utils

GetLine

hGetLineUtf8 :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es (Either UnicodeException Text)

hGetLine and decodeUtf8.

Since: 0.1

hGetLineUtf8Lenient :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Text

hGetLine and decodeUtf8Lenient.

Since: 0.1

hGetLineUtf8ThrowM :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Text

hGetLine and decodeUtf8ThrowM.

Since: 0.1

GetContents

hGetContentsUtf8 :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es (Either UnicodeException Text)

hGetContents and decodeUtf8.

Since: 0.1

hGetContentsUtf8Lenient :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Text

hGetContents and decodeUtf8Lenient.

Since: 0.1

hGetContentsUtf8ThrowM :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Eff es Text

hGetContents and decodeUtf8ThrowM.

Since: 0.1

Get

hGetUtf8 :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Int -> Eff es (Either UnicodeException Text)

hGet and decodeUtf8.

Since: 0.1

hGetUtf8Lenient :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Int -> Eff es Text

hGet and decodeUtf8Lenient.

Since: 0.1

hGetUtf8ThrowM :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Int -> Eff es Text

hGet and decodeUtf8ThrowM.

Since: 0.1

GetSome

hGetSomeUtf8 :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Int -> Eff es (Either UnicodeException Text)

hGetSome and decodeUtf8.

Since: 0.1

hGetSomeUtf8Lenient :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Int -> Eff es Text

hGetSome and decodeUtf8Lenient.

Since: 0.1

hGetSomeUtf8ThrowM :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Int -> Eff es Text

hGetSome and decodeUtf8ThrowM.

Since: 0.1

GetNonBlocking

hGetNonBlockingUtf8 :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Int -> Eff es (Either UnicodeException Text)

hGetNonBlocking and decodeUtf8.

Since: 0.1

hGetNonBlockingUtf8Lenient :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Int -> Eff es Text

hGetNonBlocking and decodeUtf8Lenient.

Since: 0.1

hGetNonBlockingUtf8ThrowM :: forall (p :: HandleMode) (es :: [Effect]). (CanRead p, HandleReader :> es, HasCallStack) => Handle p -> Int -> Eff es Text

hGetNonBlocking and decodeUtf8ThrowM.

Since: 0.1

Re-exports

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Char8 it can be interpreted as containing 8-bit characters.

type OsPath = OsString

Type representing filenames/pathnames.

This type doesn't add any guarantees over OsString.

data Text

A space efficient, packed, unboxed Unicode text type.

data UnicodeException

An exception type for representing Unicode encoding errors.