Effectful.FileSystem.PathWriter.Utils

Description

Provides utilities used by path writing.

Since: 0.1

Config

data CopyDirConfig

Directory copying config.

Since: 0.1

Constructors

MkCopyDirConfig

Fields

overwrite :: !Overwrite

Overwrite behavior.

Since: 0.1

targetName :: !TargetName

TargetName behavior.

Since: 0.1

defaultCopyDirConfig :: CopyDirConfig

Default config for copying directories.

>>> defaultCopyDirConfig
MkCopyDirConfig {overwrite = OverwriteNone, destName = TargetNameSrc}

Since: 0.1

data Overwrite

Determines file/directory overwrite behavior.

Since: 0.1

Constructors

OverwriteNone
No overwriting allowed.

Since: 0.1

OverwriteDirectories
Allow overwriting directories.

Since: 0.1

OverwriteAll
Allow overwriting the target directory and all subpaths.

Since: 0.1

data TargetName

Determines how to name the target.

Since: 0.1

Constructors

TargetNameSrc
Uses the src dir as the dest name i.e. dest/<src>.

Since: 0.1

TargetNameLiteral !OsPath
Uses the given literal as the dest name i.e. dest/<targetName>.

Since: 0.1

TargetNameDest
Uses dest itself as the target i.e. dest/ (top-level copy).

Since: 0.1

Optics

_OverwriteNone :: Prism' Overwrite ()

Since: 0.1

_OverwriteDirectories :: Prism' Overwrite ()

Since: 0.1

_OverwriteAll :: Prism' Overwrite ()

Since: 0.1

_TargetNameSrc :: Prism' TargetName ()

Since: 0.1

_TargetNameLiteral :: Prism' TargetName OsPath

Since: 0.1

_TargetNameDest :: Prism' TargetName ()

Since: 0.1