Effectful.Haskeline.Static

Description

Provides a static effect for haskeline.

Since: 0.1

Effect

data Haskeline (a :: Type -> Type) b

Static haskeline effect.

Since: 0.1

haveTerminalUI :: forall (es :: [Effect]). (HasCallStack, Haskeline :> es, Reader (InputTEnv IO) :> es) => Eff es Bool

Lifted haveTerminalUI.

Since: 0.1

getInputLine :: forall (es :: [Effect]). (HasCallStack, Haskeline :> es, Reader (InputTEnv IO) :> es) => String -> Eff es (Maybe String)

Lifted getInputLine.

Since: 0.1

getInputLineWithInitial :: forall (es :: [Effect]). (HasCallStack, Haskeline :> es, Reader (InputTEnv IO) :> es) => String -> (String, String) -> Eff es (Maybe String)

Lifted getInputLineWithInitial.

Since: 0.1

getInputChar :: forall (es :: [Effect]). (HasCallStack, Haskeline :> es, Reader (InputTEnv IO) :> es) => String -> Eff es (Maybe Char)

Lifted getInputChar.

Since: 0.1

getPassword :: forall (es :: [Effect]). (HasCallStack, Haskeline :> es, Reader (InputTEnv IO) :> es) => Maybe Char -> String -> Eff es (Maybe String)

Lifted getPassword.

Since: 0.1

waitForAnyKey :: forall (es :: [Effect]). (HasCallStack, Haskeline :> es, Reader (InputTEnv IO) :> es) => String -> Eff es Bool

Lifted waitForAnyKey.

Since: 0.1

outputStr :: forall (es :: [Effect]). (HasCallStack, Haskeline :> es, Reader (InputTEnv IO) :> es) => String -> Eff es ()

Lifted outputStr.

Since: 0.1

outputStrLn :: forall (es :: [Effect]). (HasCallStack, Haskeline :> es, Reader (InputTEnv IO) :> es) => String -> Eff es ()

Lifted outputStrLn.

Since: 0.1

getHistory :: forall (es :: [Effect]). (HasCallStack, Haskeline :> es, Reader (InputTEnv IO) :> es) => Eff es History

Lifted getHistory.

Since: 0.1

putHistory :: forall (es :: [Effect]). (HasCallStack, Haskeline :> es, Reader (InputTEnv IO) :> es) => History -> Eff es ()

Lifted putHistory.

Since: 0.1

modifyHistory :: forall (es :: [Effect]). (HasCallStack, Haskeline :> es, Reader (InputTEnv IO) :> es) => (History -> History) -> Eff es ()

Lifted modifyHistory.

Since: 0.1

withInterrupt :: forall (es :: [Effect]) a. (HasCallStack, Haskeline :> es, Reader (InputTEnv IO) :> es) => Eff es a -> Eff es a

Lifted withInterrupt.

Since: 0.1

handleInterrupt :: forall (es :: [Effect]) a. (HasCallStack, Haskeline :> es, Reader (InputTEnv IO) :> es) => Eff es a -> Eff es a -> Eff es a

Lifted handleInterrupt.

Since: 0.1

Handlers

runHaskeline :: forall (es :: [Effect]) a. (HasCallStack, IOE :> es) => Eff (Haskeline ': es) a -> Eff es a

Runs Haskeline in IO.

Since: 0.1

Reader

These functions allow eliminating the Haskeline effect in terms of ReaderT.

Examples:

 app :: (Haskeline :> es, Reader (InputTEnv IO) :> es) => Eff es String
 app = do
   mLine <- getInputLine "Enter your name: "
   let name = fromMaybe "<blank>" mLine
   outputStrLn $ "Hello: " ++ name
   pure name

 appIO :: IO String
 appIO = do
   runInputTEnv
     $ env -> runEffInputTEnv env
     $ runHaskeline app

runEffInputTEnv

Arguments

:: forall (m :: Type -> Type) a. HasCallStack

=> InputTEnv m
The InputT environment.

-> Eff '[Reader (InputTEnv m), IOE] a
Eff action that requires the InputTEnv environment.

-> IO a
IO result.

Runner for Eff with InputTEnv. Intended for usage with runInputTEnv or runInputTEnvWith.

Since: 0.1

runInputTEnv

Arguments

:: (HasCallStack, MonadIO m, MonadMask m)

=> (InputTEnv m -> m a)
Action.

-> m a
IO result.

runInputTEnvWith with default haskeline settings.

Examples:

-- eff :: Eff [Reader (InputTEnv m), IOE] a
runInputTEnv $ env -> runEffInputTEnv env eff

Since: 0.1

runInputTEnvWith

Arguments

:: HasCallStack

=> (InputT m a -> m a)
InputT runner.

-> (InputTEnv m -> m a)
Action.

-> m a
IO result.

Runs ReaderT InputTEnv in IO with InputT runner.

Examples:

-- eff :: Eff [Reader (InputTEnv m), IOE] a
runInputTEnvWith runInput $ env -> runEffInputTEnv env eff

Since: 0.1

Haskeline Re-exports

Types

data InputT (m :: Type -> Type) a

A monad transformer which carries all of the state and settings relevant to a line-reading application.

data InputTEnv (m :: Type -> Type)

The abstract environment used by InputT, for ReaderT usage.

Since: haskeline-0.8.4.0

IO Runners

runInputT :: (MonadIO m, MonadMask m) => Settings m -> InputT m a -> m a

Run a line-reading application. This function should suffice for most applications.

This function is equivalent to runInputTBehavior defaultBehavior. It uses terminal-style interaction if stdin is connected to a terminal and has echoing enabled. Otherwise (e.g., if stdin is a pipe), it uses file-style interaction.

If it uses terminal-style interaction, Prefs will be read from the user's ~/.haskeline file (if present). If it uses file-style interaction, Prefs are not relevant and will not be read.

runInputTBehavior :: (MonadIO m, MonadMask m) => Behavior -> Settings m -> InputT m a -> m a

Run a line-reading application according to the given behavior.

If it uses terminal-style interaction, Prefs will be read from the user's ~/.haskeline file (if present). If it uses file-style interaction, Prefs are not relevant and will not be read.

runInputTBehaviorWithPrefs :: (MonadIO m, MonadMask m) => Behavior -> Prefs -> Settings m -> InputT m a -> m a

Run a line-reading application.

Config

defaultSettings :: forall (m :: Type -> Type). MonadIO m => Settings m

A useful default. In particular:

defaultSettings = Settings {
          complete = completeFilename,
          historyFile = Nothing,
          autoAddHistory = True
          }

defaultBehavior :: Behavior

Read input from stdin. Use terminal-style interaction if stdin is connected to a terminal and has echoing enabled. Otherwise (e.g., if stdin is a pipe), use file-style interaction.

This behavior should suffice for most applications.

defaultPrefs :: Prefs

The default preferences which may be overwritten in the .haskeline file.