Effectful.Logger.Dynamic

Description

Dynamic effect for Control.Monad.Logger.

Since: 0.1

Effect

data Logger (a :: Type -> Type) b where

Dynamic logging effect for Control.Monad.Logger.

Since: 0.1

Constructors

LoggerLog :: forall msg (a :: Type -> Type). ToLogStr msg => Loc -> LogSource -> LogLevel -> msg -> Logger a ()

loggerLog :: forall (es :: [Effect]) msg. (HasCallStack, Logger :> es, ToLogStr msg) => Loc -> LogSource -> LogLevel -> msg -> Eff es ()

Writes a log.

Since: 0.1

Types

data LogLevel

Since: 0.1

Constructors

LevelTrace
Since: 0.1

LevelDebug
Since: 0.1

LevelInfo
Since: 0.1

LevelWarn
Since: 0.1

LevelError
Since: 0.1

LevelFatal
Since: 0.1

LevelOther Text
Since: 0.1

type LogLine = (Loc, LogSource, LogLevel, LogStr)

Since: 0.1

type LogSource = Text

Since: 0.1

Re-export from fast-logger

data LogStr

Log message builder. Use (<>) to append two LogStr in O(1).

class ToLogStr msg where

Types that can be converted to a LogStr. Instances for types from the text library use a UTF-8 encoding. Instances for numerical types use a decimal encoding.

Methods

toLogStr :: msg -> LogStr

fromLogStr :: LogStr -> ByteString

Converting LogStr to ByteString.

TH logging

logTrace :: Q Exp

See logDebug

Since: 0.1

logDebug :: Q Exp

Generates a function that takes a Text and logs a LevelDebug message. Usage:

$(logDebug) "This is a debug log message"

Since: 0.1

logInfo :: Q Exp

See logDebug

Since: 0.1

logWarn :: Q Exp

See logDebug

Since: 0.1

logError :: Q Exp

See logDebug

Since: 0.1

logFatal :: Q Exp

See logDebug

Since: 0.1

logOther :: Text -> Q Exp

Generates a function that takes a Text and logs a LevelOther message. Usage:

$(logOther "My new level") "This is a log message"

Since: 0.1

TH logging of showable values

logTraceSH :: Q Exp

See logDebugSH

Since: 0.1

logDebugSH :: Q Exp

Generates a function that takes a 'Show a => a' and logs a LevelDebug message. Usage:

$(logDebugSH) (Just "This is a debug log message")

Since: 0.1

logInfoSH :: Q Exp

See logDebugSH

Since: 0.1

logWarnSH :: Q Exp

See logDebugSH

Since: 0.1

logErrorSH :: Q Exp

See logDebugSH

Since: 0.1

logFatalSH :: Q Exp

See logDebugSH

Since: 0.1

logOtherSH :: Text -> Q Exp

Generates a function that takes a 'Show a => a' and logs a LevelOther message. Usage:

$(logOtherSH "My new level") "This is a log message"

Since: 0.1

TH logging with source

logTraceS :: Q Exp

See logDebugS

Since: 0.1

logDebugS :: Q Exp

Generates a function that takes a LogSource and Text and logs a LevelDebug message. Usage:

$logDebugS "SomeSource" "This is a debug log message"

Since: 0.1

logInfoS :: Q Exp

See logDebugS

Since: 0.1

logWarnS :: Q Exp

See logDebugS

Since: 0.1

logErrorS :: Q Exp

See logDebugS

Since: 0.1

logFatalS :: Q Exp

See logDebugS

Since: 0.1

logOtherS :: Q Exp

Generates a function that takes a LogSource, a level name and a Text and logs a LevelOther message. Usage:

$logOtherS "SomeSource" "My new level" "This is a log message"

Since: 0.1

TH util

liftLoc :: Loc -> Q Exp

Lift a location into an Exp.

Since: 0.1

Non-TH logging

logTraceN :: forall (es :: [Effect]). Logger :> es => Text -> Eff es ()

Since: 0.1

logDebugN :: forall (es :: [Effect]). Logger :> es => Text -> Eff es ()

Since: 0.1

logInfoN :: forall (es :: [Effect]). Logger :> es => Text -> Eff es ()

Since: 0.1

logWarnN :: forall (es :: [Effect]). Logger :> es => Text -> Eff es ()

Since: 0.1

logErrorN :: forall (es :: [Effect]). Logger :> es => Text -> Eff es ()

Since: 0.1

logFatalN :: forall (es :: [Effect]). Logger :> es => Text -> Eff es ()

Since: 0.1

logOtherN :: forall (es :: [Effect]). Logger :> es => LogLevel -> Text -> Eff es ()

Since: 0.1

Non-TH logging with source

logWithoutLoc :: forall (es :: [Effect]) msg. (HasCallStack, Logger :> es, ToLogStr msg) => LogSource -> LogLevel -> msg -> Eff es ()

Since: 0.1

logTraceNS :: forall (es :: [Effect]). Logger :> es => LogSource -> Text -> Eff es ()

Since: 0.1

logDebugNS :: forall (es :: [Effect]). Logger :> es => LogSource -> Text -> Eff es ()

Since: 0.1

logInfoNS :: forall (es :: [Effect]). Logger :> es => LogSource -> Text -> Eff es ()

Since: 0.1

logWarnNS :: forall (es :: [Effect]). Logger :> es => LogSource -> Text -> Eff es ()

Since: 0.1

logErrorNS :: forall (es :: [Effect]). Logger :> es => LogSource -> Text -> Eff es ()

Since: 0.1

logFatalNS :: forall (es :: [Effect]). Logger :> es => LogSource -> Text -> Eff es ()

Since: 0.1

logOtherNS :: forall (es :: [Effect]). (HasCallStack, Logger :> es) => LogSource -> LogLevel -> Text -> Eff es ()

Since: 0.1

Callstack logging

logTraceCS :: forall (es :: [Effect]). (HasCallStack, Logger :> es) => CallStack -> Text -> Eff es ()

See logDebugCS

Since: 0.1

logDebugCS :: forall (es :: [Effect]). (HasCallStack, Logger :> es) => CallStack -> Text -> Eff es ()

Logs a message with location given by CallStack. See CallStack for more convenient functions for CallStack based logging.

Since: 0.1

logInfoCS :: forall (es :: [Effect]). (HasCallStack, Logger :> es) => CallStack -> Text -> Eff es ()

See logDebugCS

Since: 0.1

logWarnCS :: forall (es :: [Effect]). (HasCallStack, Logger :> es) => CallStack -> Text -> Eff es ()

See logDebugCS

Since: 0.1

logErrorCS :: forall (es :: [Effect]). (HasCallStack, Logger :> es) => CallStack -> Text -> Eff es ()

See logDebugCS

Since: 0.1

logFatalCS :: forall (es :: [Effect]). (HasCallStack, Logger :> es) => CallStack -> Text -> Eff es ()

See logDebugCS

Since: 0.1

logOtherCS :: forall (es :: [Effect]). (HasCallStack, Logger :> es) => CallStack -> LogLevel -> Text -> Eff es ()

See logDebugCS

Since: 0.1

utilities for defining your own loggers

defaultLogStr :: Loc -> LogSource -> LogLevel -> LogStr -> LogStr

Since: 0.1

data Loc

A location within a source file.

Constructors

Loc

Fields

loc_filename :: String

loc_package :: String

loc_module :: String

loc_start :: CharPos

loc_end :: CharPos

defaultLoc :: Loc

dummy location, used with logWithoutLoc

Since: 0.1

defaultOutput :: HasCallStack => Handle -> Loc -> LogSource -> LogLevel -> LogStr -> IO ()

A default implementation of loggerLog that accepts a file handle as the first argument.

This is used in the definition of runStdoutLoggingT:

runStdoutLoggingT :: MonadIO m => LoggingT m a -> m a
runStdoutLoggingT action =
    runLoggingT action (defaultOutput stdout)

Since: 0.1

Level checks

guardLevel

Arguments

:: (Applicative f, HasCallStack)

=> LogLevel
The configured log level to check against.

-> LogLevel
The log level for this action.

-> (HasCallStack => f ())
The logging action to run if the level passes.

-> f ()

guardLevel configLvl lvl m runs m iff shouldLog configLvl lvl.

Since: 0.1

shouldLog

Arguments

:: LogLevel
The configured log level to check against.

-> LogLevel
Level for this log

-> Bool
Whether we should log

shouldLog configLvl lvl returns true iff configLvl <= lvl. Uses LogLevel's built-in ordering. The ordering is thus:

  LevelTrace
    < LevelDebug
    < LevelInfo
    < LevelWarn
    < LevelError
    < LevelFatal
    < LevelOther "<any>"

In other words, LogLevel's usual Ord is respected. Note that LevelOther "custom" sits at the the highest level and compares via Text's Ord, just like LogLevel's usual Ord.

Since: 0.1

Formatting

data LogFormatter

Formatter for logs.

Since: 0.1

Constructors

MkLogFormatter

Fields

locStrategy :: !LocStrategy

How to log the code location.

Since: 0.1

newline :: !Bool

If true, append a newline.

Since: 0.1

threadLabel :: !Bool

Whether to include the thread's label set by labelThread. Falls back to the thread's ThreadId when the label has not been set.

Since: 0.1

timezone :: !Bool

Whether to include the timezone in the timestamp.

Since: 0.1

defaultLogFormatter :: Loc -> LogFormatter

LogFormatter with:

locStrategy = LocPartial loc
newline = True
threadLabel = False
timezone = False

Since: 0.1

data LocStrategy

Determines how we log location data.

Since: 0.1

Constructors

LocPartial !Loc
Logs the location with filename, line, col.

Since: 0.1

LocStable !Loc
Logs the location with filename.

Since: 0.1

LocNone
No location logging.

Since: 0.1

formatLog

Arguments

:: forall msg (es :: [Effect]). (Concurrent :> es, HasCallStack, Time :> es, ToLogStr msg)

=> LogFormatter
Formatter to use.

-> LogLevel
The level in which to log.

-> msg
Message.

-> Eff es LogStr
Formatted LogStr.

Produces a formatted LogStr in terms of:

Static Concurrent.

Dynamic Time.

Example

-- [timestamp][thread_label][code_loc][level] msg
[2022-02-08 10:20:05][thread-label][filename:1:2][Warn] msg

Since: 0.1

Optics

_LevelTrace :: Prism' LogLevel ()

Since: 0.1

_LevelInfo :: Prism' LogLevel ()

Since: 0.1

_LevelDebug :: Prism' LogLevel ()

Since: 0.1

_LevelWarn :: Prism' LogLevel ()

Since: 0.1

_LevelError :: Prism' LogLevel ()

Since: 0.1

_LevelOther :: Prism' LogLevel Text

Since: 0.1

_LevelFatal :: Prism' LogLevel ()

Since: 0.1