Effectful.Notify.Static

Description

Provides an effect for desktop notifications.

Since: 0.1

Description

effects-notify is a library for implementing cross-platform desktop notifications. Currently supports:

Linux: Supports dbus and notify-send. That is, if there is a running dbus notification server, it can be used directly. notify-send (libnotify) can also be used, though it also requires a running dbus server.

Osx: Supports built-in apple-script i.e. osascript.

Windows: Unsupported. Functions are mere stubs that will compile (with a warning) and do nothing at runtime.

Usage

General usage involves choosing the notification system and using that to initialize the environment:

-- choose system
let systemOs :: NotifySystemOs
    systemOs = ...

-- prepare environment
notifyEnv <- initNotifyEnv systemOs

-- send notification
let note =
      mkNote "A summary"
        & setBody "Some notification"
        & setTitle "My Application"

notify notifyEnv note

Choosing the system is more interesting. If we do not care about being cross-platform, we can simply choose a NotifySystemOs directly and pass it to init:

-- The NotifySystemOsDBus constructor is only available on linux.
let systemOs :: NotifySystemOs
    systemOs = NotifySystemOsDBus

notifyEnv <- initNotifyEnv systemOs
...

If we need to work on several platforms, we can instead use NotifySystem, which includes every possible notification system and is available on all platforms. We then pass it to notifySystemToOs, which -- if successful -- will return the NotifySystemOs and we can continue as normal.

-- All NotifySystem constructors are available on all platforms.
let system :: NotifySystem
    system = ...

systemOs <- notifySystemToOs system >>= \case
  Left ex -> throwM ex
  Right s -> pure s
notifyEnv <- initNotifyEnv systemOs
...

If a system is chosen that is unavailable on the current platform, a NotifyParseException will be returned.

Effect

data Notify (a :: Type -> Type) b

Static notify effect.

Since: 0.1

initNotifyEnv

Arguments

:: forall (es :: [Effect]). (HasCallStack, Notify :> es)

=> NotifySystemOs
System.

-> Eff es NotifyEnv

Initialize the notification environment.

Since: 0.1

notify

Arguments

:: forall (es :: [Effect]). (HasCallStack, Notify :> es)

=> NotifyEnv
Env.

-> Note
Note.

-> Eff es ()

Send a notification.

Since: 0.1

Handlers

runNotify :: forall (es :: [Effect]) a. (HasCallStack, IOE :> es) => Eff (Notify ': es) a -> Eff es a

Runs a Notify effect in IO.

Since: 0.1

Errors

newtype NotifyInitException

Exceptions thrown when initializing notifications.

Since: 0.1

Constructors

MkNotifyInitException

Fields

unNotifyInitException :: SomeException

data NotifyException

Exceptions thrown when sending notifications. The list of non-fatal exceptions includes:

DBus: Non-fatal client errors.

All others are considered fatal.

Since: 0.1

Constructors

MkNotifyException

Fields

exception :: SomeException

Underlying exception.

fatal :: Bool

Is fatal?

note :: Note

Note we attempted to send.

notifySystem :: NotifySystem

Notification system.

catchNonFatalNotify

Arguments

:: forall (es :: [Effect]). (HasCallStack, Notify :> es)

=> NotifyEnv
Env.

-> Note
Note.

-> (NotifyException -> Eff es ())
Handler.

-> Eff es ()

Runs notify and runs the handler on any non-fatal NotifyExceptions. Any other exceptions are rethrown.

Since: 0.1

tryNonFatalNotify

Arguments

:: forall (es :: [Effect]). (HasCallStack, Notify :> es)

=> NotifyEnv
Env.

-> Note
Note.

-> Eff es (Maybe NotifyException)

Like notify, but catches non-fatal NotifyException.

Since: 0.1

tryNonFatalNotify_

Arguments

:: forall (es :: [Effect]). (HasCallStack, Notify :> es)

=> NotifyEnv
Env.

-> Note
Note.

-> Eff es ()

Like tryNonFatalNotify, but catches and discards the exception.

Since: 0.1

Types

System

Platform-specific

data NotifySystemOs

Notification system. Options are platform specific.

Since: 0.1

Constructors

NotifySystemOsDBus

NotifySystemOsNotifySend

defaultNotifySystemOs :: NotifySystemOs

Since: 0.1

All platforms

data NotifySystem

Represents all possible notification systems. This is useful for when we want to choose the system in some way that works across multiple platforms without requiring cpp.

Since: 0.1

Constructors

NotifySystemAppleScript

NotifySystemDBus

NotifySystemNotifySend

NotifySystemWindows

defaultNotifySystem :: NotifySystem

Since: 0.1

notifySystemToOs :: NotifySystem -> Either NotifyParseException NotifySystemOs

Converts platform-agnostic NotifySystem to platform-specific NotifySystemOs. Useful for obtaining a valid system for the current platform without cpp.

Since: 0.1

notifySystemFromOs :: NotifySystemOs -> NotifySystem

Inverse notifySystemToOs.

Since: 0.1

data Os

Os types.

Since: 0.1

Constructors

Linux
Since: 0.1

Osx
Since: 0.1

Windows
Since: 0.1

data NotifyParseException

Exceptions thrown when parsing NotifySystem to NotifySystemOs.

Since: 0.1

Constructors

MkNotifyParseException

Fields

os :: Os

The current Os.

system :: NotifySystem

The NotifySystem we attempted to use.

Env

data NotifyEnv

Notification environment.

Since: 0.1

notifyEnvToSystemOs :: NotifyEnv -> NotifySystemOs

Gives the system for the given environment.

Since: 0.1

Notes

data Note

Notification type.

Since: 0.1

mkNote :: Text -> Note

Makes a note from a text summary.

setBody :: Maybe Text -> Note -> Note

Since: 0.1

getBody :: Note -> Maybe Text

Since: 0.1

setSummary :: Text -> Note -> Note

Since: 0.1

getSummary :: Note -> Text

Since: 0.1

setTimeout :: Maybe NotifyTimeout -> Note -> Note

Since: 0.1

getTimeout :: Note -> Maybe NotifyTimeout

Since: 0.1

setTitle :: Maybe Text -> Note -> Note

Since: 0.1

getTitle :: Note -> Maybe Text

Since: 0.1

setUrgency :: Maybe NotifyUrgency -> Note -> Note

Since: 0.1

getUrgency :: Note -> Maybe NotifyUrgency

Since: 0.1

Timeout

data NotifyTimeout

Notification timeout, in milliseconds.

Since: 0.1

Constructors

NotifyTimeoutNever

NotifyTimeoutMillis Int

_NotifyTimeoutMillis :: Prism' NotifyTimeout Int

Since: 0.1

_NotifyTimeoutNever :: Prism' NotifyTimeout ()

Since: 0.1

NotifyUrgency

data NotifyUrgency

Notification urgency. Not supported by all notification systems.

Constructors

NotifyUrgencyLow
Low.

NotifyUrgencyNormal
Normal.

NotifyUrgencyCritical
Critical.

_NotifyUrgencyLow :: Prism' NotifyUrgency ()

Since: 0.1

_NotifyUrgencyNormal :: Prism' NotifyUrgency ()

Since: 0.1

_NotifyUrgencyCritical :: Prism' NotifyUrgency ()

Since: 0.1