Effectful.Posix.Files.Static

Description

Provides a static effect for System.Posix.Files.

Since: 0.1

Effect

data PosixFiles (a :: Type -> Type) b

Provides a static effect for System.Posix.Files.

Since: 0.1

setFileMode :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> FileMode -> Eff es ()

Lifted setFileMode.

Since: 0.1

setFdMode :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => Fd -> FileMode -> Eff es ()

Lifted setFdMode.

Since: 0.1

setFileCreationMask :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => FileMode -> Eff es FileMode

Lifted setFileCreationMask.

Since: 0.1

fileAccess :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> Bool -> Bool -> Bool -> Eff es Bool

Lifted fileAccess.

Since: 0.1

fileExist :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> Eff es Bool

Lifted fileExist.

Since: 0.1

getFileStatus :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> Eff es FileStatus

Lifted getFileStatus.

Since: 0.1

getFdStatus :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => Fd -> Eff es FileStatus

Lifted getFdStatus.

Since: 0.1

getSymbolicLinkStatus :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> Eff es FileStatus

Lifted getSymbolicLinkStatus.

Since: 0.1

createNamedPipe :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> FileMode -> Eff es ()

Lifted createNamedPipe.

Since: 0.1

createDevice :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> FileMode -> DeviceID -> Eff es ()

Lifted createDevice.

Since: 0.1

createLink :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> PosixPath -> Eff es ()

Lifted createLink.

Since: 0.1

removeLink :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> Eff es ()

Lifted removeLink.

Since: 0.1

createSymbolicLink :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> PosixPath -> Eff es ()

Lifted createSymbolicLink.

Since: 0.1

readSymbolicLink :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> Eff es PosixPath

Lifted readSymbolicLink.

Since: 0.1

rename :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> PosixPath -> Eff es ()

Lifted rename.

Since: 0.1

setOwnerAndGroup :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> UserID -> GroupID -> Eff es ()

Lifted setOwnerAndGroup.

Since: 0.1

setFdOwnerAndGroup :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => Fd -> UserID -> GroupID -> Eff es ()

Lifted setFdOwnerAndGroup.

Since: 0.1

setSymbolicLinkOwnerAndGroup :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> UserID -> GroupID -> Eff es ()

Lifted setSymbolicLinkOwnerAndGroup.

Since: 0.1

setFileTimes :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> EpochTime -> EpochTime -> Eff es ()

Lifted setFileTimes.

Since: 0.1

touchFile :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> Eff es ()

Lifted touchFile.

Since: 0.1

setFileSize :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> FileOffset -> Eff es ()

Lifted setFileSize.

Since: 0.1

setFdSize :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => Fd -> FileOffset -> Eff es ()

Lifted setFdSize.

Since: 0.1

getPathVar :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> PathVar -> Eff es Limit

Lifted getPathVar.

Since: 0.1

getFdPathVar :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => Fd -> PathVar -> Eff es Limit

Lifted getFdPathVar.

Since: 0.1

Handler

runPosixFiles :: forall (es :: [Effect]) a. (HasCallStack, IOE :> es) => Eff (PosixFiles ': es) a -> Eff es a

Runs a PosixFiles effect.

Since: 0.1

PathType

data PathType

Path type.

Since: fs-utils-0.1

Constructors

PathTypeFile
Since: fs-utils-0.1

PathTypeDirectory
Since: fs-utils-0.1

PathTypeSymbolicLink
Since: fs-utils-0.1

PathTypeOther
Since: fs-utils-0.1

displayPathType :: IsString a => PathType -> a

String representation of PathType.

Since: fs-utils-0.1

Functions

throwIfWrongPathType :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => String -> PathType -> PosixPath -> Eff es ()

Throws IOException if the path does not exist or the expected path type does not match actual.

Since: 0.1

isPathType :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PathType -> PosixPath -> Eff es Bool

Checks that the path type matches the expectation. Throws IOException if the path does not exist or the type cannot be detected.

Since: 0.1

getPathType :: forall (es :: [Effect]). (HasCallStack, PosixFiles :> es) => PosixPath -> Eff es PathType

Returns the type for a given path without following symlinks. Throws IOException if the path does not exist or the type cannot be detected.

Since: 0.1