Effectful.Time.Dynamic

Description

Provides a dynamic effect for Data.Time.

Since: 0.1

Effect

data Time (a :: Type -> Type) b where

Dynamic effect for Data.Time.

Since: 0.1

Constructors

GetSystemZonedTime :: forall (a :: Type -> Type). Time a ZonedTime

GetTimeZone :: forall (a :: Type -> Type). UTCTime -> Time a TimeZone

UtcToLocalZonedTime :: forall (a :: Type -> Type). UTCTime -> Time a ZonedTime

LoadLocalTZ :: forall (a :: Type -> Type). Time a TZ

GetMonotonicTime :: forall (a :: Type -> Type). Time a Double

getSystemTime :: forall (es :: [Effect]). (HasCallStack, Time :> es) => Eff es LocalTime

Returns the local system time.

Since: 0.1

getSystemZonedTime :: forall (es :: [Effect]). (HasCallStack, Time :> es) => Eff es ZonedTime

Returns the zoned system time.

Since: 0.1

getTimeZone :: forall (es :: [Effect]). (HasCallStack, Time :> es) => UTCTime -> Eff es TimeZone

Lifted getTimeZone.

Since: 0.1

utcToLocalZonedTime :: forall (es :: [Effect]). (HasCallStack, Time :> es) => UTCTime -> Eff es ZonedTime

Lifted utcToLocalZonedTime.

Since: 0.1

loadLocalTZ :: forall (es :: [Effect]). (HasCallStack, Time :> es) => Eff es TZ

Lifted loadLocalTZ.

Since: 0.1

getMonotonicTime :: forall (es :: [Effect]). (HasCallStack, Time :> es) => Eff es Double

Returns the zoned system time

Since: 0.1

Handlers

runTime :: forall (es :: [Effect]) a. (HasCallStack, IOE :> es) => Eff (Time ': es) a -> Eff es a

Runs Time in IO.

Since: 0.1

Timing

withTiming :: forall (es :: [Effect]) a. (HasCallStack, Time :> es) => Eff es a -> Eff es (TimeSpec, a)

Runs an action, returning the elapsed time.

Since: 0.1

withTiming_ :: forall (es :: [Effect]) a. (HasCallStack, Time :> es) => Eff es a -> Eff es TimeSpec

withTiming but ignores the result value.

Since: 0.1

TimeSpec

data TimeSpec

Structure for holding time data. Eq and Ord are defined in terms of an equivalence class e.g.

MkTimeSpec s n === MkTimeSpec 0 (s * 1_000_000_000 + n)

Since: 0.1

Constructors

MkTimeSpec

Fields

sec :: !Natural

Seconds.

Since: 0.1

nsec :: !Natural

Nanoseconds.

Since: 0.1

Creation

fromSeconds :: Double -> TimeSpec

Converts a Double to a TimeSpec.

Since: 0.1

fromNanoSeconds :: Natural -> TimeSpec

Converts Natural nanoseconds to a TimeSpec.

Since: 0.1

Elimination

toSeconds :: TimeSpec -> Double

Converts a TimeSpec to a Double.

Since: 0.1

toNanoSeconds :: TimeSpec -> Natural

Converts a TimeSpec into Natural nanoseconds.

Since: 0.1

Operations

diffTimeSpec :: TimeSpec -> TimeSpec -> TimeSpec

Returns the absolute difference of two TimeSpecs.

Since: 0.1

normalizeTimeSpec :: TimeSpec -> TimeSpec

Normalizes nanoseconds < 1 second.

Since: 0.1

Formatting

formatLocalTime :: LocalTime -> String

Formats the LocalTime to YYYY-MM-DD HH:MM:SS.

Since: 0.1

formatZonedTime :: ZonedTime -> String

Formats the ZonedTime to YYYY-MM-DD HH:MM:SS Z.

Since: 0.1

Parsing

parseLocalTime :: (HasCallStack, MonadFail f) => String -> f LocalTime

Parses the LocalTime from YYYY-MM-DD HH:MM:SS.

Since: 0.1

parseZonedTime :: (HasCallStack, MonadFail f) => String -> f ZonedTime

Parses the ZonedTime from YYYY-MM-DD HH:MM:SS Z.

Known Timezones

Expand

UTC

UT

GMT

EST

EDT

CST

CDT

MST

MDT

PST

PDT

+HHMM (e.g. +1300)

Since: 0.1

Misc

getSystemTimeString :: forall (es :: [Effect]). (HasCallStack, Time :> es) => Eff es String

Retrieves the formatted LocalTime.

Since: 0.1

getSystemZonedTimeString :: forall (es :: [Effect]). (HasCallStack, Time :> es) => Eff es String

Retrieves the formatted ZonedTime.

Since: 0.1

Re-exports

data LocalTime

A simple day and time aggregate, where the day is of the specified parameter, and the time is a TimeOfDay. Conversion of this (as local civil time) to UTC depends on the time zone. Conversion of this (as local mean time) to UT1 depends on the longitude.

Constructors

LocalTime

Fields

localDay :: Day

localTimeOfDay :: TimeOfDay

data ZonedTime

A local time together with a time zone.

There is no Eq instance for ZonedTime. If you want to compare local times, use zonedTimeToLocalTime. If you want to compare absolute times, use zonedTimeToUTC.

Constructors

ZonedTime

Fields

zonedTimeToLocalTime :: LocalTime

zonedTimeZone :: TimeZone